added feign Client and added event delete functionality

+++ src/main/java/com/veersa/eventApp/DTO/PaymentResponse.java
@@ -12,5 +12,5 @@ public class PaymentResponse {
     private  String paymentLink;
     private  String message;
     private  String paymentId;
-    private  Integer ammount;
+    private  Double ammount;
 }
